Rename testHelpers to fit convention
