+++ client/src/components/Upload/Composite.test.js
@@ -1,5 +1,5 @@
 import Composite from "./Composite.vue";
-import { mountWithApp } from "./test_helpers";
+import { mountWithApp } from "./testHelpers";
 
 describe("Composite.vue", () => {
     it("loads with correct initial state", async () => {
