+++ client/src/components/Upload/RulesInput.test.js
@@ -1,5 +1,5 @@
 import RulesInput from "./RulesInput.vue";
-import { mountWithApp } from "./test_helpers";
+import { mountWithApp } from "./testHelpers";
 
 describe("RulesInput.vue", () => {
     it("loads with correct initial state", async () => {
